consolidate CreateDefaultConfigurationSource() into a new ConfigurationSource helper
